Move "liquid-fire-transitions" service implementation into "addon" folder
